+++ docs/changelog.md
@@ -26,6 +26,7 @@ Unreleased.
    lists and sets inadvertently included a trailing space.
  * Comments are now allowed in more places, in particular right after `=>` in
    functions.
+ * Add [`Set.to_list`](type_set.md#to_list) method.
 
 ## 0.11.0